don't allocate a new vec in every `ChunkVecBuffer::write_to`
This commit changes the implementation of `ChunkVecBuffer::write_to` so
that a vector is not allocated in every call. This should hopefully
reduce the overhead of writing TLS data to a stream.

Instead of using a `Vec`, a fixed-size array of 64 `IoSlice`s is used.
If the number of queued chunks exceeds 64, we may write fewer bytes in
one `write_to` call, but not allocating and dropping a vec in every call
is probably worth it. My intuition is that 64 `IoSlice`s is plenty and
it shouldn't be exceeded often, but I'll admit that it was chosen fairly
arbitrarily.

Another potential approach could be to use a single vec that lives in
the `ChunkVecBuffer`, and is `clear`ed when consuming written data,
retaining the allocated capacity. That way, the existing allocation is
reused, and we only allocate when the `VecDeque` of queued chunks has
grown longer than its the previous maximum length. However, this is not
presently possible, because an `IoSlice` is borrowed, and we can't add a
`Vec<IoSlice<'a>>` to the `ChunkVecBuffer`, since the lifetime `'a` here
would be the lifetime of the `&mut self` receiver of `write_to` --- so
the slices would not live long enough to be assigned to such a vec.

Therefore, I think using a stack array is the right approach for now.