Bump MSRV to Rust 1.47.0.
Rust 1.47.0 gets rid of `LengthAtMost32`. This allows *ring* to expose a
better API in `ring::rand` and to have a more clearly correct
implementation with less `unsafe`. Rust 1.46.0 added support for more
expressions in `const fn`s which will let me clean up some more parts of
*ring* to make them more clearly correct. So, I'm eager to require at
least Rust 1.47.0 for *ring*. I'd like to update Rustls to use the
latest version of *ring* that I plan to release this month.

Rustls PR #428 would be cleaner if we support at least Rust 1.44; see
the comment in the PR for details.