Simplify the handling of SCT and OCSP sending in the server.
If we're not going to send one of these, then drop it as early as
possible to reduce memory overhead, so that at the time of sending,
`CertifiedKey` will only contain it if it is to be sent. This makes
the boolean flags that previously controlled this redundant, so
remove them.

In the process, remove some unwrapping.