Ensure SNI extension can be round-tripped
In 6940729 this was made to use webpki::DNSName as the underlying
storage.  However, that type case smashes to lower case.

This means a handshake where a upper case character is in the SNI
extension fails: we accidentally pretend that the hostname was
all lower case, and the client/server transcripts become out of sync.

Keep SNI hostnames twice in memory: first the raw bytes (used for encoding),
second the validated DNSName.