Remove clone in certificate receipt
Rename require_handshake_msg_mut to require_handshake_msg_move:
it's not really used for mutating incoming messages, but does
partially move them.

This it the problem with using it everywhere: some messages have
to be processed prior to being added to the transcript.  Moving part
of them prevents this.