Ticket: Reduce locking overhead and eliminate some panic potential.
Before: Lock. Maybe roll. Unlock. Lock. Do the work. Unlock.
After:  Lock. Maybe roll.               Do the work. Unlock.

This is still not ideal because the lock is held too long, and held over
inappropriate operations (syscalls). I will file a separate issue about
that. It was the case before this change as well.