Avoid panicking in the unlikely case that the time is set very wrong.
If the system time is set to before the Unix epoch (probably not possible on
most operating systems) then computing the duration since the Unix epoch will
fail. Avoid that very unlikely case.

Add the time of retrieval (storage) to `ClientSessionValueWithResolvedCipherSuite`
since the lifetime of the needed time is exactly the lifetime of the value of that
type. (Perhaps `ClientSessionValueWithResolvedCipherSuite` could be given a better
name.)

In the future, if we were to refactor the `Ticekter` API to accept the current time
then we could probably reduce the number of times the current time is retrieved by
adding a `TimeBase` parameter to each of `encrypt()` and `decrypt()`.

Also stengthen the type used to represent the current time from `u64` to a new
`TimeBase` type.