Reduce Github token permissions to read-only access.
Follow the advice near the end of
https://github.blog/changelog/2021-04-20-github-actions-control-permissions-for-github_token/.

By default the GitHub token has read-write permissions, and actions may use the
GitHub token without it being explicitly mentioned in Rustls's YAML. (For example,
action/checkout@v2 uses the GitHub token.) The consequence of this is that a buggy
or maliciously-modified (third-party) action may change the contents of the Rustls
repository, e.g. insert back doors or remove security-critical checks and tests.

The repository owner can change the default to read-only but this isn't something
that others can see. So, even though ctz may have (and should) change that default,
also reset the permissions in the YAML so that outside observers do not need to
rely on him having done that.