Track clean connection closure
The only way to see an Ok(0) from Connection::read() is via a close_notify alert.

Unclean closure (ie, just a TCP EOF without a close_notify) leads
to an `UnexpectedEof` from Connection::read().  Applications which
want to accept that can do so -- bogo_shim is such an example.

Add tests for this and IoState::peer_has_closed()