chore: remove trytls shim and helper script.
The `trytls/runme` script has bitrot and doesn't function on systems
with up to date Python installations. The upstream project has Python
3.8+ compatibility issues and hasn't seen new commits in ~7 years.

Rustls has easier to maintain test coverage through
`examples/tests/badssl.rs` and the bogo test suite. For that reason this
commit removes the `trytls` scaffolding in favour of continuing to
maintain it.