verify: remove unneeded deprecated.
The `client_auth_root_subjects` fn doesn't rely on a deprecated function
any longer, and so we can remove the explicit allow.

Similarly, while previously the `sig` field of the
`DigitallySignedStruct` was marked deprecated (to push users towards an
accessor method). We've removed the `pub` visibility and the deprecation
so the `encode` method no longer needs to allow deprecated usages.