tests: add API test for Stream read unexpected EOF.
This commit introduces a unit tests that verify that a `Stream` or
`StreamOwned` that receives an EOF from a peer that hasn't signalled the
intent to close should read an `UnexpectedEOF` error.

Additionally, since there was no existing coverage for `read_buf` this
is added along with some small helpers.