Fix connect-tests for 1000-sans.badssl.com
This regressed in ae4ca32.  DNS names starting with
and wholly numerals are in wide use.

RFC1035 actually doesn't really reflect reality, and RFC1123
updates it to allow labels that start with numerals or are
wholly numeric.  The only restriction is that the last label
cannot be wholly numeric.

Also:

- Limit the entire length of a DNS name.
- Ensure a label that ends with hyphens respects the label length limit.